#3 - Security management

+++ src/AppBundle/Controller/TaskController.php
@@ -30,7 +30,7 @@ class TaskController extends Controller
 
         if ($form->isValid()) {
             $em = $this->getDoctrine()->getManager();
-            $comment->setUser($this->getUser())->setDateCreated(new \DateTime());
+            $task->setUser($this->getUser());
             $em->persist($task);
             $em->flush();
 
@@ -83,12 +83,15 @@ class TaskController extends Controller
      */
     public function deleteTaskAction(Task $task)
     {
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($task);
-        $em->flush();
+        if (($task->getUser() === $this->getUser()) || (empty($task->getUser() && $this->isGranted('ROLE_ADMIN')))) {
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($task);
+            $em->flush();
 
-        $this->addFlash('success', 'La tâche a bien été supprimée.');
-
+            $this->addFlash('success', 'La tâche a bien été supprimée.');
+        } else {
+            $this->addFlash('error', 'Vous ne pouvez pas supprimer cette tâche.');
+        }
         return $this->redirectToRoute('task_list');
     }
 }
